Minor refactor of protocol list.
