+++ app/javascript/components/admin_app/components/ProtocolDetailCard.jsx
@@ -6,7 +6,7 @@ import {CopyToClipboard} from "react-copy-to-clipboard";
 import BootstrapRoutesButton from "bootstrap-switch-button-react";
 import {ProtocolDefinitionType, StudyDefinitionType} from "../../../types";
 import PropTypes from "prop-types";
-import {connect, useSelector} from "react-redux";
+import {connect} from "react-redux";
 
 class _Protocol extends React.Component {
   constructor(props) {
@@ -60,7 +60,7 @@ class _Protocol extends React.Component {
             <CopyToClipboard
               text={`${window.location.origin}/studies/${this.props.study.id}/protocols/${this.props.protocol.id}`}
               onCopy={() => this.setState({copied: true})}>
-              <button type="button" className="btn btn-primary">
+              <button type="button" className="btn btn-primary protocol-link-button">
                 {this.state.copied ? <span style={{color: 'red'}}>Copied.</span> : <span>Get Link</span>}
                 &nbsp;
                 <i className="fas fa-link" aria-hidden="true"/>
@@ -113,30 +113,3 @@ const mapDispatchToProps = (dispatch) => ({
 
 export const EditableProtocolCard = connect(_state => ({}), mapDispatchToProps)(_Protocol);
 
-export function EditableProtocolList(props) {
-  const protocols = useSelector(state => state.protocol_definitions)
-
-  let protocol_list = props.study_protocols.map((protocol, _i) => {
-    // This is a little gross.  Because the protocol_definitions inside the study definitions
-    // don't get updated when we patch the protocol definition, we need to check if there's
-    // a protocol_definition in the protocol_definitions state which matches the id, and treat
-    // that as authoritative.
-    let protocol_def = protocols.data.filter((p) => (p.id === protocol.id));
-    if (protocol_def && (protocol_def.length > 0)) {
-      protocol = protocol_def[0]
-    }
-    return <EditableProtocolCard protocol={protocol} study={props.study} key={protocol.id}/>;
-  });
-
-  return (
-    <div className="row study-info-row" key={"new-protocol"}>
-      <b className="col-2">Protocols:</b>
-      <div className="col-10">{protocol_list}</div>
-    </div>
-  );
-}
-
-EditableProtocolList.propTypes = {
-  study_protocols: PropTypes.arrayOf(ProtocolDefinitionType),
-  study: StudyDefinitionType,
-};
